refactor: moved helpers to services.py and validators to validators.py
